Actualizacion controlador base

+++ servidor/controladores/ControladorClubs.js
@@ -16,7 +16,6 @@ class ControladorClubs extends ControladorBase {
             CARPETA: {
                 CARPETA: CARPETA,
                 CAMPO: CAMPO,
-                nombreFoto: 'principal.jpg'
             },
             TABLA: TABLA,
             QUERIES: QueriesClub,
@@ -31,28 +30,31 @@ class ControladorClubs extends ControladorBase {
             method
         } = req.route.stack[0];
         //       if (method.toLowerCase() === "put") { //<- luego, repetimos para insert y delete....
-        let id = req.params.id;
-        let sqlDelete = 'DELETE  FROM club_music WHERE clubid = ?';
-        let sqlInsert = 'INSERT INTO club_music (clubid, musicid) VALUES  '
-        let musicsUpdates = req.body.musicsUpdate.split(',');
-        super.sendDataToTable([id], sqlDelete)
-            .then(result1 => console.log("borrado", result1))
-            .then(() => {
-                if (method.toLowerCase() !== "delete") {
-                    let salida = sqlInsert;
-                    musicsUpdates.map(value => {
-                        salida += "(" + id + "," + value + "),";
-                    });
-                    salida = salida.substring(0, salida.length - 1)
-                    console.log('sql: ', salida);
-                    super.sendDataToTable([], salida) //es promesa
-                        .then(result2 => console.log("ESTE ES EL INSERT", result2))
-                        .catch(err => console.error(err));
-                }
-            })
+        if (method.toLowerCase() === "put") {
+
+            let id = req.params.id;
+            let sqlDelete = 'DELETE  FROM club_music WHERE clubid = ?';
+            let sqlInsert = 'INSERT INTO club_music (clubid, musicid) VALUES  '
+            let musicsUpdates = req.body.musicsUpdate.split(',');
+            super.sendDataToTable([id], sqlDelete)
+                .then(result1 => console.log("borrado", result1))
+                .then(() => {
+                    if (method.toLowerCase() !== "delete") {
+                        let salida = sqlInsert;
+                        musicsUpdates.map(value => {
+                            salida += "(" + id + "," + value + "),";
+                        });
+                        salida = salida.substring(0, salida.length - 1)
+                        console.log('sql: ', salida);
+                        super.sendDataToTable([], salida) //es promesa
+                            .then(result2 => console.log("ESTE ES EL INSERT", result2))
+                            .catch(err => console.error(err));
+                    }
+                })
 
 
-        delete req.body['musicsUpdate'];
+            delete req.body['musicsUpdate'];
+        }
         super.updateTable(req, res);
     }
